first take: FirehoseWriter, consumer cmd, and Dockerfile
Much of this is copied or built upon the SampleRecordProcessor from
https://github.com/clever/amazon-kinesis-client-go

The `firehose` package is based on our existing heka-clever-plugin
https://github.com/Clever/heka-clever-plugins/blob/master/firehose_output.go
and uses the same `batcher` to batch messages before sending.

Note: The Dockerfile is currently not very optimized (it sets up Go
internally and does the build there, moreover it runs the commands from
the Makefile). This was the quickest way for me to bootstrap, but should
be improved later.