writer: checkpoint only flushed messages
Previously, the checkpointing logic could checkpoint *ahead* of the
messages being sent to Firehose (since it happened when processing
records, instead of after sending the batch). Now, we pass along
sequence numbers to the batcher in order to track this more accurately.

The writer now tracks the `lastSequenceNumberFlushed`, and sets this
value when checkpointing.

The goal here is to make sure we deliver messages "at least once".